test: test citgm not just bin
